python-server-dashboard/dashboard-web-python/requirements/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
Bootstrap-Flask==2.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dominate==2.8.0
Flask==2.3.2
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multi-key-dict==2.0.3
numpy==1.25.2
pandas==2.1.0
paramiko==3.3.1
pbr==5.11.1
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-gitlab==3.15.0
python-jenkins==1.8.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
requests-toolbelt==1.0.0
schedule==1.2.0
six==1.16.0
soupsieve==2.5
subprocess.run==0.0.8
sync-to-async==0.1.1
tzdata==2023.3
urllib3==1.26.16
visitor==0.1.3
Werkzeug==2.3.7
WTForms==3.0.1
